Activate virtual environment within pulse.py (#343)
